requirements.txt
flask==1.1.2
uWSGI==2.0.18
pyyaml==5.3.1
flask_swagger_ui==3.25.0
azure-keyvault==4.1.0
azure-cli==2.5.1
azure-identity==1.3.1
azure-storage-file==2.1.0
azure.cosmos==4.0.0
azure.mgmt.privatedns==0.1.0
applicationinsights==0.11.9
rootpath==0.1.1
flask_api==2.0
tenacity==6.2.0
mock==3.0.5
more_itertools==8.2.0
